Rename `block.new`
